imagelab-backend/pyproject.toml
[project]
name = "imagelab-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.0",
    "opencv-python-headless>=4.10",
    "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "pre-commit>=4.0",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html:htmlcov -q"

[tool.coverage.run]
source = ["app"]
omit = ["app/database.py", "app/__init__.py", "app/operators/__init__.py", "app/operators/basic/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
